UI of operation pump with builder and parser of main request/response message from/to pump controller done.

+++ TouchGFX/gui/include/gui/inputsalescreen_screen/InputSaleScreenView.hpp
@@ -16,7 +16,7 @@ public:
 
 protected:
 
-    uint32_t liters;
+    float liters;
 };
 
 #endif // INPUTSALESCREENVIEW_HPP
